Add an option to add additional forwarders to ForwardedHeadersMiddleware.
  - not all webservers seem to use the default ForwardedHeaders but use their own
