+++ src/Microsoft.AspNetCore.HttpOverrides/Forwarder.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.AspNetCore.Http;
+
+namespace Microsoft.AspNetCore.Builder
+{
+    public abstract class Forwarder
+    {
+        public abstract void ApplyForwarders(HttpContext context);
+    }
+}
